# Add table aws_costoptimizationhub_recommendation Closes #2350