--- a/aws/plugin.go
+++ b/aws/plugin.go
@@ -179,6 +179,7 @@ func Plugin(ctx context.Context) *plugin.Plugin {
 			"aws_cost_by_tag":                                              tableAwsCostByTag(ctx),
 			"aws_cost_forecast_daily":                                      tableAwsCostForecastDaily(ctx),
 			"aws_cost_forecast_monthly":                                    tableAwsCostForecastMonthly(ctx),
+			"aws_costoptimizationhub_recommendation":                       tableAwsCostOptimizationHubRecommendation(ctx),
 			"aws_cost_usage":                                               tableAwsCostAndUsage(ctx),
 			"aws_dax_cluster":                                              tableAwsDaxCluster(ctx),
 			"aws_dax_parameter":                                            tableAwsDaxParameter(ctx),