--- a/aws/table_aws_costoptimizationhub_recommendation.go
+++ b/aws/table_aws_costoptimizationhub_recommendation.go
@@ -227,7 +227,6 @@ func listCostOptimizationHubRecommendations(ctx context.Context, d *plugin.Query
 		limit := int32(*d.QueryContext.Limit)
 		if limit < maxLimit {
 			maxLimit = limit
-			maxLimit = limit
 		}
 	}