@misraved 2024-12-17
@ParthaI should the table name be `aws_cost_optimization_hub_recommendation`?