@ParthaI 2024-12-17
@misraved, I named it following the Terraform naming convention, which we typically use as our first choice. Reference: [Terraform Cost Optimization Hub Preferences](https://registry.terraform.io/providers/hashicorp/aws/latest/docs/resources/costoptimizationhub_preferences).

Do you think we still need to rename it?